% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_v2.R
\name{Boxcox}
\alias{Boxcox}
\title{Box-Cox Translataion}
\usage{
Boxcox(y, lambda = 1)
}
\arguments{
\item{y}{input Time-seires data}

\item{lambda}{lamdba ( default = 1)}
}
\value{
Output - Box-Cox Translataion
}
\description{
\code{Boxcox} Box-Cox Translataion
}
\examples{
Boxcox(AirPassengers )
}
