% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_v2.R
\name{lag_xreg}
\alias{lag_xreg}
\title{The make xreg variable  for modeling}
\usage{
lag_xreg(x, maxlag)
}
\arguments{
\item{maxlag}{make data variable}

\item{y}{input External Variable data}
}
\value{
Translate External Variable data for modeling
}
\description{
\code{lag_y} xreg data translate for modeling
}
\examples{
lag_xreg(AirPassengers, 5)
}
