% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_v2.R
\name{lag_y}
\alias{lag_y}
\title{The make variable for modeling}
\usage{
lag_y(y, maxlag)
}
\arguments{
\item{y}{input time-seires data}

\item{maxlag}{make data variable}
}
\value{
Translate time-seires data for modeling
}
\description{
\code{lag_y} TimeSeries data translate for modeling
}
\examples{
lag_y(AirPassengers, 5)
}
