% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xg_v2.R
\name{xgboost.forecast}
\alias{xgboost.forecast}
\title{The analysis using xgboost}
\usage{
xgboost.forecast(y, xreg = NULL, pred_xreg = NULL, h = 42,
  maxlag = max(42, 4 * frequency(y)), nrounds = 500, params = NULL,
  nrounds_method = c("cv", "none"), nfold = ifelse(length(y) > 720, 10,
  5), boxcox = TRUE, lambda = BoxCox.lambda(y), ts_clean = TRUE,
  ts_lambda = 1, season_type = c("decompose", "dummy", "none"),
  verbose = TRUE, ...)
}
\arguments{
\item{y}{input time-seires vector}

\item{xreg}{External Variable for using modeling (need to matrix class)}

\item{pred_xreg}{External Variable for using forecast (need to matrix class )}

\item{h}{forecast preiod}

\item{maxlag}{data lag length}

\item{nrounds_method}{cv : Cross-Vaildation / none - not using}

\item{boxcox}{boxcox (TRUE / FALSE)}

\item{ts_clean}{clean Outlier (TRUE / FALSE)}

\item{ts_lambda}{lambda with clean Outlier}

\item{season_type}{treat a Type of season ('dummy', "decompose", "none")}
}
\value{
time-seires data forecast
}
\description{
\code{xgboost.forecast} Analysis/Forecast Using xgboost
}
\examples{
xgboost.forecast(AirPassengers)
}
