% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_v2.R
\name{cat.forecast}
\alias{cat.forecast}
\title{The analysis using Catboost}
\usage{
cat.forecast(y, h = 42, xreg = NULL, pred_xreg = NULL,
  params = NULL, maxlag = max(42, 4 * frequency(y)), valid = TRUE,
  boxcox = TRUE, ts_clean = TRUE, ts_lambda = 1,
  season_type = c("dummy", "decompose", "none"), verbose = TRUE, ...)
}
\arguments{
\item{y}{input time-seires vector}

\item{h}{forecast preiod}

\item{xreg}{External Variable for using modeling (need to matrix class)}

\item{pred_xreg}{External Variable for using forecast (need to matrix class )}

\item{maxlag}{data lag length}

\item{valid}{Validation (TRUE / FALSE)}

\item{boxcox}{boxcox (TRUE / FALSE)}

\item{ts_clean}{clean Outlier (TRUE / FALSE)}

\item{ts_lambda}{lambda with clean Outlier}

\item{season_type}{treat a Type of season ('dummy', "decompose", "none")}
}
\value{
time-seires data forecast
}
\description{
\code{cat.forecast} Analysis/Forecast Using Catboost
}
\examples{
cat.forecast(AirPassengers)
}
